update code to use nftables set
Originally the idea of this program was to simply list the IP
networks belonging to a set of countries so that they could
be used to update nftables through `nft`.

One of the main problems with the above was that the list were
too long.

Since nftables already uses intervals, I thought to use them
too, with the added bonus that I could join subnets and
end up with fewer elements.

Since this is still quite experimental, I used an hack that, so
far only works for IPv4.